portfolio-backend/requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.0
Pillow>=10.0
drf-spectacular>=0.27
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.0
dj-database-url>=2.0
psycopg2-binary>=2.9
